+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## 0.0.2
+
+- Breaking change: `FirebaseAdmin.initializeApp()` now accepts separate named
+    arguments for credential and databaseURL
+- Breaking change: `Credential.cert()` now accepts separate named arguments
+    for service account key paramenters
+- New `Credential.certFromPath()` method added.
+- Added `DataSnapshot.forEach()`.
+- Many dartdoc changes
+
 ## 0.0.1
 
 - Initial version